Remove define_ephemeral_datastore macro.
Instead, move the code inside the macro directly into the
datastore::test_util module, and update other uses of the macro to
refer to this version.
